ObjectFileable is merged into ObjectFile
