+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2022-02-28 21:11:02 UTC using RuboCop version 1.25.1.
+# on 2022-06-14 23:06:50 UTC using RuboCop version 1.30.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -17,10 +17,15 @@ Metrics/AbcSize:
   Max: 55
 
 # Offense count: 1
-# Configuration parameters: CountComments, CountAsOne, ExcludedMethods, IgnoredMethods.
+# Configuration parameters: CountComments, CountAsOne, ExcludedMethods, IgnoredMethods, inherit_mode.
 # IgnoredMethods: refine
 Metrics/BlockLength:
   Max: 27
+
+# Offense count: 1
+# Configuration parameters: CountComments, CountAsOne.
+Metrics/ClassLength:
+  Max: 130
 
 # Offense count: 2
 # Configuration parameters: IgnoredMethods.
@@ -31,11 +36,6 @@ Metrics/CyclomaticComplexity:
 # Configuration parameters: CountComments, CountAsOne, ExcludedMethods, IgnoredMethods.
 Metrics/MethodLength:
   Max: 31
-
-# Offense count: 1
-# Configuration parameters: CountComments, CountAsOne.
-Metrics/ModuleLength:
-  Max: 120
 
 # Offense count: 1
 # Configuration parameters: CountKeywordArgs, MaxOptionalParameters.
@@ -64,7 +64,7 @@ Naming/FileName:
 Naming/PredicateName:
   Exclude:
     - 'spec/**/*'
-    - 'lib/assembly-objectfile/object_fileable.rb'
+    - 'lib/assembly-objectfile/object_file.rb'
 
 # Offense count: 1
 # Configuration parameters: Prefixes.
@@ -86,7 +86,7 @@ RSpec/FilePath:
     - 'spec/content_metadata_spec.rb'
     - 'spec/object_file_spec.rb'
 
-# Offense count: 74
+# Offense count: 3
 # Configuration parameters: AssignmentOnly.
 RSpec/InstanceVariable:
   Exclude:
@@ -116,14 +116,14 @@ RSpec/RepeatedExampleGroupDescription:
     - 'spec/content_metadata_spec.rb'
 
 # Offense count: 2
-# Cop supports --auto-correct.
+# This cop supports safe autocorrection (--autocorrect).
 Style/CommentedKeyword:
   Exclude:
     - 'lib/assembly-objectfile/content_metadata.rb'
 
-# Offense count: 123
-# Cop supports --auto-correct.
-# Configuration parameters: AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, IgnoredPatterns.
+# Offense count: 122
+# This cop supports safe autocorrection (--autocorrect).
+# Configuration parameters: AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, AllowedPatterns, IgnoredPatterns.
 # URISchemes: http, https
 Layout/LineLength:
   Max: 277
